+++ torchmetrics/functional/classification/confusion_matrix.py
@@ -14,15 +14,14 @@
 from typing import Optional
 
 import torch
+from torch import Tensor
 
-from torchmetrics.classification.checks import _input_format_classification
 from torchmetrics.utilities import rank_zero_warn
+from torchmetrics.utilities.checks import _input_format_classification
 from torchmetrics.utilities.enums import DataType
 
 
-def _confusion_matrix_update(
-    preds: torch.Tensor, target: torch.Tensor, num_classes: int, threshold: float = 0.5
-) -> torch.Tensor:
+def _confusion_matrix_update(preds: Tensor, target: Tensor, num_classes: int, threshold: float = 0.5) -> Tensor:
     preds, target, mode = _input_format_classification(preds, target, threshold)
     if mode not in (DataType.BINARY, DataType.MULTILABEL):
         preds = preds.argmax(dim=1)
@@ -33,12 +32,13 @@ def _confusion_matrix_update(
     return confmat
 
 
-def _confusion_matrix_compute(confmat: torch.Tensor, normalize: Optional[str] = None) -> torch.Tensor:
+def _confusion_matrix_compute(confmat: Tensor, normalize: Optional[str] = None) -> Tensor:
     allowed_normalize = ('true', 'pred', 'all', 'none', None)
     assert normalize in allowed_normalize, \
         f"Argument average needs to one of the following: {allowed_normalize}"
     confmat = confmat.float()
     if normalize is not None and normalize != 'none':
+        cm = None
         if normalize == 'true':
             cm = confmat / confmat.sum(axis=1, keepdim=True)
         elif normalize == 'pred':
@@ -54,12 +54,8 @@ def _confusion_matrix_compute(confmat: torch.Tensor, normalize: Optional[str] = 
 
 
 def confusion_matrix(
-    preds: torch.Tensor,
-    target: torch.Tensor,
-    num_classes: int,
-    normalize: Optional[str] = None,
-    threshold: float = 0.5
-) -> torch.Tensor:
+    preds: Tensor, target: Tensor, num_classes: int, normalize: Optional[str] = None, threshold: float = 0.5
+) -> Tensor:
     """
     Computes the confusion matrix. Works with binary, multiclass, and multilabel data.
     Accepts probabilities from a model output or integer class values in prediction.
@@ -86,7 +82,6 @@ def confusion_matrix(
             Threshold value for binary or multi-label probabilities. default: 0.5
 
     Example:
-
         >>> from torchmetrics.functional import confusion_matrix
         >>> target = torch.tensor([1, 1, 0, 0])
         >>> preds = torch.tensor([0, 1, 0, 0])